+++ src/Module/Widget/GetAddress/Html.php
@@ -42,7 +42,7 @@ class Html extends Widget
         public readonly string $btnClassList = ''
     ) {
         $this->content = $this->render(
-            file: __DIR__ . DIRECTORY_SEPARATOR . 'templates' .
+            file: $this->getWidgetName() . DIRECTORY_SEPARATOR . 'templates' .
             DIRECTORY_SEPARATOR . 'html.phtml'
         );
     }