+++ src/Module/Widget/GetAddress/Js.php
@@ -46,7 +46,7 @@ class Js extends Widget
         $stringValidation->isUrl(value: $url);
 
         $this->content = $this->render(
-            file: __DIR__ . DIRECTORY_SEPARATOR . 'templates' .
+            file: $this->getWidgetName() . DIRECTORY_SEPARATOR . 'templates' .
             DIRECTORY_SEPARATOR . 'js.js.phtml'
         );
     }