+++ src/Module/Widget/PartPayment/Js.php
@@ -90,7 +90,7 @@ class Js extends Widget
         );
 
         $this->content = $this->render(
-            file: __DIR__ . DIRECTORY_SEPARATOR . 'templates' .
+            file: $this->getWidgetName() . DIRECTORY_SEPARATOR . 'templates' .
             DIRECTORY_SEPARATOR . 'js.js.phtml'
         );
     }