+++ src/Module/Widget/ReadMore/Html.php
@@ -76,7 +76,7 @@ class Html extends Widget
             $this->getLegacyUrl() : $this->getSecciUrl();
 
         $this->content = $this->render(
-            file: __DIR__ . DIRECTORY_SEPARATOR . 'templates' .
+            file: $this->getWidgetName() . DIRECTORY_SEPARATOR . 'templates' .
             DIRECTORY_SEPARATOR . 'html.phtml'
         );
     }