+++ src/Module/Widget/UniqueSellingPoint/Html.php
@@ -71,7 +71,7 @@ class Html extends Widget
             useLegacyLink: $this->useLegacyReadMoreLink
         );
         $this->content = $this->render(
-            file: __DIR__ . DIRECTORY_SEPARATOR . 'templates' .
+            file: $this->getWidgetName() . DIRECTORY_SEPARATOR . 'templates' .
             DIRECTORY_SEPARATOR . 'html.phtml'
         );
     }