+++ tests/Integration/Module/Payment/Api/RefundTest.php
@@ -21,6 +21,7 @@ use Resursbank\Ecom\Exception\AttributeCombinationException;
 use Resursbank\Ecom\Exception\AuthException;
 use Resursbank\Ecom\Exception\ConfigException;
 use Resursbank\Ecom\Exception\CurlException;
+use Resursbank\Ecom\Exception\TimeoutException;
 use Resursbank\Ecom\Exception\Validation\EmptyValueException;
 use Resursbank\Ecom\Exception\Validation\IllegalTypeException;
 use Resursbank\Ecom\Exception\Validation\IllegalValueException;
@@ -156,7 +157,17 @@ class RefundTest extends TestCase
         $payment = $this->createPayment(orderReference: $orderReference);
 
         // Sign
-        MockSigner::callCustomerUrl(payment: $payment);
+        try {
+            MockSigner::callCustomerUrl(payment: $payment);
+        } catch (TimeoutException $error) {
+            if ($_ENV['IS_PIPELINE']) {
+                $this->markTestSkipped(
+                    message: 'MockSigner failed with timeout.'
+                );
+            }
+
+            throw $error;
+        }
 
         // Capture payment
         Repository::capture(paymentId: $payment->id);
@@ -198,7 +209,17 @@ class RefundTest extends TestCase
         $payment = $this->createPayment(orderReference: $orderReference);
 
         // Sign
-        MockSigner::callCustomerUrl(payment: $payment);
+        try {
+            MockSigner::callCustomerUrl(payment: $payment);
+        } catch (TimeoutException $error) {
+            if ($_ENV['IS_PIPELINE']) {
+                $this->markTestSkipped(
+                    message: 'MockSigner failed with timeout.'
+                );
+            }
+
+            throw $error;
+        }
 
         // Capture
         Repository::capture(paymentId: $payment->id);
@@ -261,7 +282,17 @@ class RefundTest extends TestCase
         $payment = $this->createPayment(orderReference: $orderReference);
 
         // Sign
-        MockSigner::callCustomerUrl(payment: $payment);
+        try {
+            MockSigner::callCustomerUrl(payment: $payment);
+        } catch (TimeoutException $error) {
+            if ($_ENV['IS_PIPELINE']) {
+                $this->markTestSkipped(
+                    message: 'MockSigner failed with timeout.'
+                );
+            }
+
+            throw $error;
+        }
 
         // Capture
         Repository::capture(paymentId: $payment->id);
@@ -315,7 +346,17 @@ class RefundTest extends TestCase
         $payment = $this->createPayment(orderReference: $orderReference);
 
         // Sign
-        MockSigner::callCustomerUrl(payment: $payment);
+        try {
+            MockSigner::callCustomerUrl(payment: $payment);
+        } catch (TimeoutException $error) {
+            if ($_ENV['IS_PIPELINE']) {
+                $this->markTestSkipped(
+                    message: 'MockSigner failed with timeout.'
+                );
+            }
+
+            throw $error;
+        }
 
         // Capture
         Repository::capture(paymentId: $payment->id);
@@ -368,7 +409,17 @@ class RefundTest extends TestCase
         $payment = $this->createPayment(orderReference: $orderReference);
 
         // Sign
-        MockSigner::callCustomerUrl(payment: $payment);
+        try {
+            MockSigner::callCustomerUrl(payment: $payment);
+        } catch (TimeoutException $error) {
+            if ($_ENV['IS_PIPELINE']) {
+                $this->markTestSkipped(
+                    message: 'MockSigner failed with timeout.'
+                );
+            }
+
+            throw $error;
+        }
 
         // Capture
         Repository::capture(paymentId: $payment->id);